refactor: split the add_properties family out and factor three oversized fns
Lint and structural work, kept out of the feature commits so the sprint's
substance stays reviewable. No behaviour changes; every suite green before and
after.

`maintain.rs` was 2609 lines at this branch's base — already past the 2500
ceiling — and the Part 2 index fix nudged it to 2630. Per the standing rule the
file is split rather than the ceiling raised: the `add_properties` family moves
to `mutation/add_properties.rs` (one public entry point, its aggregate branch,
and the expression/geometry helpers only those two use). maintain.rs drops to
2043 and the new module is 611, so the inherited breach is resolved rather than
merely not worsened. `AddPropertiesReport` deliberately stays defined in
maintain.rs: it is the return type of the public
`kglite::api::mutation::add_properties`, and the pinned Rust API baseline records
it at that canonical path, so moving the definition would churn a public
signature for an internal reorganisation.

Three functions that gained a match arm in this sprint are factored instead of
having their recorded metrics raised:

- `execute_call` 787→704 lines, 132→100 branches. Its ~40-arm YIELD-column table
  had nothing to do with executing a call and grew with every new procedure; it
  is now `valid_yield_columns`. The `db.indexes` / `db.constraints` dispatch arms
  were byte-identical and are merged.
- `write_type_detail` 406→364 lines, 47→40 branches, nesting 6→5. The
  per-property `<prop .../>` attribute assembly is now `property_attrs`.
- `execute_single_clause` back to its recorded 329/47. The schema read/mutation
  split moves into `schema_ddl::is_schema_read` / `execute_schema_read`, next to
  both implementations, so the routing arm is one case and `clause_is_mutation`
  has a single place to agree with.

Every source-quality baseline edit therefore moves **downward**; no ceiling was
raised, no file allowance added, no entry created. The only non-numeric change is
`add_properties_aggregate`'s recorded path following the split.

Clippy `--all-targets -D warnings` is clean on kglite, kglite-py, and
kglite-mcp-server. One error was this branch's (a helper landed after a test
module in describe.rs); the other five were inherited and are fixed here since
they blocked the stack: four `needless_borrows_for_generic_args` on
`check_required_fields` / `unique_claims` call sites, and a
`From<ConstraintViolation>` impl after the test module in error.rs.
`set_schema` gets `#[allow(clippy::result_large_err)]`, matching the convention
already used at ten sites for the same lint.
