+++ crates/kglite/src/error.rs
@@ -605,6 +605,34 @@ pub type KgResult<T> = std::result::Result<T, KgError>;
 
 // ─── Tests ───────────────────────────────────────────────────────────────────
 
+impl From<crate::graph::constraints::ConstraintViolation> for KgError {
+    /// Lift a core constraint violation to the public error type, splitting on
+    /// whether a *write* or a *declaration* failed — the two carry different
+    /// Neo4j status codes and call for different fixes.
+    fn from(violation: crate::graph::constraints::ConstraintViolation) -> Self {
+        let message = violation.to_string();
+        let descriptor = violation.descriptor();
+        let kind = violation.kind.keyword();
+        if violation.is_declaration_failure() {
+            KgError::ConstraintCreationFailed {
+                kind,
+                node_type: violation.node_type,
+                properties: violation.properties,
+                descriptor,
+                message,
+            }
+        } else {
+            KgError::ConstraintViolation {
+                kind,
+                node_type: violation.node_type,
+                properties: violation.properties,
+                descriptor,
+                message,
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -739,31 +767,3 @@ mod tests {
 }
 
 // ─── Constraint bridge ───────────────────────────────────────────────────────
-
-impl From<crate::graph::constraints::ConstraintViolation> for KgError {
-    /// Lift a core constraint violation to the public error type, splitting on
-    /// whether a *write* or a *declaration* failed — the two carry different
-    /// Neo4j status codes and call for different fixes.
-    fn from(violation: crate::graph::constraints::ConstraintViolation) -> Self {
-        let message = violation.to_string();
-        let descriptor = violation.descriptor();
-        let kind = violation.kind.keyword();
-        if violation.is_declaration_failure() {
-            KgError::ConstraintCreationFailed {
-                kind,
-                node_type: violation.node_type,
-                properties: violation.properties,
-                descriptor,
-                message,
-            }
-        } else {
-            KgError::ConstraintViolation {
-                kind,
-                node_type: violation.node_type,
-                properties: violation.properties,
-                descriptor,
-                message,
-            }
-        }
-    }
-}
