+++ crates/kglite/src/graph/dir_graph/schema_ops.rs
@@ -47,6 +47,7 @@ impl DirGraph {
     /// stops enforcing it. A constraint declared directly (by
     /// `DirGraph::create_unique_constraint`, e.g. from `CREATE CONSTRAINT`)
     /// rather than through a schema is untouched by this.
+    #[allow(clippy::result_large_err)]
     pub fn set_schema(&mut self, schema: SchemaDefinition) -> Result<(), KgError> {
         let previous_schema = self.schema_definition.take();
         let withdrawn = Self::declared_unique_tuples(previous_schema.as_ref());
