+++ crates/kglite/src/lib.rs
@@ -151,10 +151,14 @@ pub mod api {
             ConnectionOperationReport, NodeOperationReport, OperationReport, OperationReports,
         };
         pub use crate::graph::mutation::extend::{extend_graph, ExtendReport};
+        // `AddPropertiesReport` is deliberately not re-exported: it was not part
+        // of the public surface before this module was split out, and the API
+        // baseline pins that surface.
+        pub use crate::graph::mutation::add_properties::{add_properties, PropertySpec};
         pub use crate::graph::mutation::maintain::{
-            add_connections, add_edges_from_specs, add_nodes, add_properties, create_connections,
+            add_connections, add_edges_from_specs, add_nodes, create_connections,
             purge_provisional_nodes, replace_connections, update_node_properties, EdgeSpec,
-            EdgeSpecReport, PropertySpec,
+            EdgeSpecReport,
         };
         /// Validate a graph against a `SchemaDefinition` (Piece 3 cleanup).
         pub use crate::graph::mutation::validation::validate_graph;
