+++ tests/api-baselines/source-quality.json
@@ -196,9 +196,9 @@
     {
       "path": "crates/kglite/src/graph/introspection/describe.rs",
       "qualified_name": "crate::write_type_detail",
-      "lines": 416,
-      "branches": 48,
-      "nesting": 6
+      "lines": 364,
+      "branches": 40,
+      "nesting": 5
     },
     {
       "path": "crates/kglite/src/graph/io/export.rs",
@@ -266,8 +266,8 @@
     {
       "path": "crates/kglite/src/graph/languages/cypher/executor/call_clause.rs",
       "qualified_name": "crate::CypherExecutor<'a>::execute_call",
-      "lines": 801,
-      "branches": 135,
+      "lines": 704,
+      "branches": 100,
       "nesting": 5
     },
     {
@@ -329,8 +329,8 @@
     {
       "path": "crates/kglite/src/graph/languages/cypher/executor/mod.rs",
       "qualified_name": "crate::CypherExecutor<'a>::execute_single_clause",
-      "lines": 332,
-      "branches": 48,
+      "lines": 329,
+      "branches": 47,
       "nesting": 3
     },
     {
@@ -558,7 +558,7 @@
       "nesting": 4
     },
     {
-      "path": "crates/kglite/src/graph/mutation/maintain.rs",
+      "path": "crates/kglite/src/graph/mutation/add_properties.rs",
       "qualified_name": "crate::add_properties_aggregate",
       "lines": 165,
       "branches": 36,
